relocated search logic from headers into search bar components
